Remove other examples, rename unifiedlog_iterator
This removes the other examples since all of the functionality is now
contained by the unifiedlog_iterator example. Also renames
unifiedlog_iterator to unifiedlog_parser
