--- examples/parse_tracev3/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "parse_tracev3"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-simplelog = "0.12.2"
-serde_json = "1.0.122"
-log = "0.4.22"
-macos-unifiedlogs = {path = "../../"}
