--- examples/unifiedlog_iterator/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "unifiedlog_iterator"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-simplelog = "0.12.2"
-csv = "1.3.0"
-chrono = "0.4.38"
-log = "0.4.22"
-serde_json = "1.0.122"
-macos-unifiedlogs = {path = "../../"}
-clap = {version = "4.5.18", features = ["derive"]}
