+++ examples/unifiedlog_parser/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "unifiedlog_parser"
+name = "unifiedlog_iterator"
 version = "0.1.0"
 edition = "2021"
 
@@ -10,5 +10,6 @@ simplelog = "0.12.2"
 csv = "1.3.0"
 chrono = "0.4.38"
 log = "0.4.22"
+serde_json = "1.0.122"
 macos-unifiedlogs = {path = "../../"}
 clap = {version = "4.5.18", features = ["derive"]}
