--- examples/unifiedlog_parser_json/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "unifiedlog_parser_json"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-simplelog = "0.12.2"
-serde_json = "1.0.122"
-log = "0.4.22"
-macos-unifiedlogs = {path = "../../"}
-clap = {version = "4.5.15", features = ["derive"]}
